+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.18.0
 
+### Added
+  
+- Latency based Cname records to aws-route53
+
 ### Changed
 
 -   Update ExtendedNodejsFunction and CdkPipeline to use Node 22 runtime as the default